Converted to Module
Updates the .ps1 file into a module file (psm1) and adds a module manifest file.
This update allows the code to more readily be deployed to a server.
Updates the README file to demonstrate that importing the module is required before use.
Renames relevant files to better demonstrate the code is now modularized.